user, journal and admin authentication and authorization added

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -1,10 +1,12 @@
 package net.engineeringdigest.journalApp.service;
 
 import net.engineeringdigest.journalApp.entity.JournalEntry;
+import net.engineeringdigest.journalApp.entity.User;
 import net.engineeringdigest.journalApp.repository.JournalEntryRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Date;
 import java.util.List;
@@ -15,8 +17,24 @@ public class JournalEntryService  {
     @Autowired
     private JournalEntryRepository journalEntryRepository;
 
+    @Autowired
+    private UserService userService;
+
+    @Transactional
+    public void saveEntry(JournalEntry journalEntry, String userName) {
+        try {
+            User user = userService.findByUserName(userName);
+            journalEntry.setDate(new Date());
+            JournalEntry saved = journalEntryRepository.save(journalEntry);
+            user.getJournalEntries().add(saved);
+            userService.saveUser(user);
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error occured in SaveEntry: ", e);
+        }
+    }
+
     public void saveEntry(JournalEntry journalEntry) {
-        journalEntry.setDate(new Date());
         journalEntryRepository.save(journalEntry);
     }
 
@@ -29,8 +47,23 @@ public class JournalEntryService  {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id) {
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("Error occured in deleting journal by id : " + e);
+        }
+        return removed;
     }
+
+
 
 }
